# backend/docs/api.yaml
openapi: 3.0.3
info:
  title: Stock Insights API
  description: |
    High-performance stock insights system that provides real-time stock data and recommendations.
    
    ## Features
    - Real-time stock data from external APIs
    - Recommendation algorithm
    - Automatic daily data ingestion
    - Advanced filtering and search capabilities
    - Job tracking and monitoring
    
    ## Authentication
    - Public endpoints: No authentication required
    - Admin endpoints: Bearer token authentication required
    
    ## Rate Limiting
    - Public endpoints: 100 requests per minute per IP
    - Admin endpoints: 50 requests per minute per IP
  version: 1.0.0
  contact:
    name: Stock Insights Team
    email: support@stock-insights.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: htts://stock-insights-production.up.railway.app
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/admin/ingest/stocks:
    post:
      summary: Trigger manual data ingestion
      description: |
        Manually trigger the stock data ingestion process.
        This will fetch new data from the external API and update the database.
        
        The ingestion process runs asynchronously and returns a job ID for tracking.
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '202':
          description: Ingestion job started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ingestion job started"
                  job_id:
                    type: string
                    format: uuid
                    example: "cc31797d-b9bc-4898-9cb4-3fef1f9beec0"
                  status:
                    type: string
                    example: "accepted"
                  job:
                    $ref: '#/components/schemas/Job'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for admin authentication.
        Include the token in the Authorization header as "Bearer <token>"
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique job ID
          example: "cc31797d-b9bc-4898-9cb4-3fef1f9beec0"
        status:
          type: string
          enum: [pending, running, completed, failed]
          description: Current job status
          example: "running"
        created_at:
          type: string
          format: date-time
          description: When the job was created
          example: "2025-08-03T07:16:29.2214745-05:00"
        started_at:
          type: string
          format: date-time
          description: When the job started executing
          example: "2025-08-03T07:16:29.2215207-05:00"
        completed_at:
          type: string
          format: date-time
          description: When the job completed (if finished)
          example: "2025-08-03T07:16:35.1234567-05:00"
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Job progress percentage
          example: 75
        message:
          type: string
          description: Current job status message
          example: "Processing stocks..."
        error:
          type: string
          description: Error message if job failed
          example: "External API timeout"
      required:
        - id
        - status
        - created_at
        - progress
        - message
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: "Unauthorized"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid or expired authentication token"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
      required:
        - error
        - message
...